remembrall/requirements.txt
asgiref==3.7.2
Django==3.2.23
django-filter==23.5
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
importlib-metadata==6.7.0
Markdown==3.4.4
packaging==23.2
phonenumbers==8.13.29
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
whitenoise==6.5.0
zipp==3.15.0
